+++ src/reduce/multiGPUstreams.h
@@ -1,9 +1,9 @@
-#include "../shared.cu.h"
+#include "../shared_cuda.cu.h"
 
 template<typename T>
 __global__ void multiGpuStreamReductionKernel(
-    const T* input_array, const T x, T* output_array, const int array_len, 
-    const int stream_num
+    const T* input_array, const T x, T* output_array, 
+    const unsigned long int array_len, const int stream_num
 ) {
     
 }
@@ -11,7 +11,8 @@ __global__ void multiGpuStreamReductionKernel(
 template<typename F, typename T>
 void multiGpuStreamReduction(
     F mapped_kernel, const T* input_array, const T constant, T* output_array, 
-    const int array_len, const cudaStream_t* streams, const int stream_count
+    const unsigned long  int array_len, const cudaStream_t* streams, 
+    const int stream_count
 ) {
     std::cout << "TODO multiGpuStreamReduction\n";
 }
